+++ modules/weko-search-ui/weko_search_ui/utils.py
@@ -3957,238 +3957,6 @@ def handle_check_filename_consistence(file_paths, meta_filenames):
     return errors
 
 
-class RoCrateConverter:
-    """Metadata to RO-Crate Converter."""
-
-    REGEX = re.compile(r'(.+)\[([0-9]+)\]$')
-    DEFAULT_LANG = 'en'
-
-    def __init__(self):
-        from rocrate.rocrate import ROCrate
-        self.crate = ROCrate()
-
-    def convert(self, metadata, format, lang='en'):
-        """
-        Convert metadata to RO-Crate.
-
-        Args:
-            metadata (dict) : record metadata
-            format (dict)   : conversion format
-            lang (str)      : language
-
-        Returns:
-            ROCrate: ROCrate class
-        """
-        from rocrate.rocrate import ROCrate
-
-        self.lang = lang
-        self.crate = ROCrate()
-
-        tree_structure = format.get('tree_structure', {})
-        entity_types = format.get('entity_types', {})
-
-        # Add file entities
-        self.__add_file_entities(metadata, format)
-
-        # Add root properties
-        root_entity = self.crate.root_dataset
-        root_map = tree_structure.get('map', {})
-        self.__add_properties(root_entity, root_map, metadata)
-
-        # Add entities
-        child_entities = self.__add_child_entities(tree_structure.get('children', {}), metadata, entity_types)
-        root_entity['hasPart'] = child_entities
-
-        return self.__get_json()
-
-    def __add_file_entities(self, metadata, format):
-        """
-        Add file entity.
-
-        Args:
-            metadata (dict) : record metadata
-            format (dict)   : conversion format
-        """
-
-        metadata_files = self.__get_metadata_files(metadata)
-        map_file = format.get('file', {}).get('map', {})
-        file_entities = []
-        for metadata_file in metadata_files:
-            file_entity = self.crate.add_file(metadata_file.get('filename', ''))
-            self.__add_properties(file_entity, map_file, metadata_file)
-            file_entities.append(file_entity)
-        self.crate.root_dataset['mainEntity'] = file_entities
-
-    def __get_metadata_files(self, metadata):
-        """
-        Get file metadata.
-
-        Args:
-            metadata (dict) : record metadata
-        """
-
-        files = []
-        for key, value in metadata.items():
-            if isinstance(value, dict) and value.get('attribute_type') == 'file':
-                files = value.get('attribute_value_mlt', [])
-                break
-
-        return files
-
-    def __add_properties(self, entity, map, metadata):
-        """
-        Add properties to entity.
-
-        Args:
-            entity          : RO-Crate entity
-            map (dict)      : mapping information
-            metadata (dict) : metadata
-        """
-
-        for property_name, item_type_key in map.items():
-            property_value = None
-            if isinstance(item_type_key, list):
-                property_value = []
-                for key in item_type_key:
-                    value = self.__get_property(key, metadata)
-                    if isinstance(value, list):
-                        property_value = property_value + value
-                    else:
-                        property_value.append(value)
-            elif isinstance(item_type_key, dict):
-                values = self.__get_property(item_type_key.get('value'), metadata)
-                languages = self.__get_property(item_type_key.get('lang'), metadata)
-                if len(values) != len(languages):
-                    continue
-                indices = [i for i, x in enumerate(languages) if x == self.lang]
-                if not indices:
-                    indices = [i for i, x in enumerate(languages) if x == self.DEFAULT_LANG]
-                if not indices:
-                    indices = [0]
-                property_value = [x for i, x in enumerate(values) if i in indices]
-            else:
-                property_value = self.__get_property(item_type_key, metadata)
-
-            if property_value:
-                entity[property_name] = property_value
-
-    def __get_property(self, target, metadata):
-        """
-        Get property value.
-
-        Args:
-            target (str)    : string with concatenated keys. (e.g., item_0000000000000.subitem_0000000000001)
-            metadata (dict) : metadata
-
-        Returns:
-            any: property value
-        """
-
-        keys = []
-        for key in target.split('.'):
-            name = None
-            index = None
-            match_result = self.REGEX.findall(key)  # Check if key has index
-            if match_result:
-                name = match_result[0][0]
-                index = int(match_result[0][1])
-            else:
-                name = key
-            keys.append({
-                "name": name,
-                "index": index,
-            })
-
-        return self.__get_property_value(keys, 0, metadata)
-
-    def __get_property_value(self, keys, depth, metadata):
-        """
-        Get property value from metadata.
-
-        Args:
-            keys (list)     : key lest divided by dot
-            depth (int)     : tree depth
-            metadata (dict) : metadata
-
-        Returns:
-            any: property value
-        """
-
-        if len(keys) <= depth:
-            return None
-
-        key_name = keys[depth].get('name')
-        index = keys[depth].get('index')
-
-        if isinstance(metadata, dict):
-            if key_name not in metadata:
-                return None
-            metadata = metadata.get(key_name)
-        if isinstance(metadata, dict):
-            if 'attribute_value_mlt' in metadata:
-                metadata = metadata.get('attribute_value_mlt')
-        if index is not None and isinstance(metadata, list):
-            if len(metadata) <= index:
-                return None
-            metadata = metadata[index]
-
-        property_value = None
-        if isinstance(metadata, dict):
-            property_value = self.__get_property_value(keys, depth + 1, metadata)
-        elif isinstance(metadata, list):
-            property_value = []
-            for metadatum in metadata:
-                property_value.append(self.__get_property_value(keys, depth + 1, metadatum))
-        else:
-            property_value = metadata
-
-        return property_value
-
-    def __add_child_entities(self, children, metadata, entity_types, parent_id=''):
-        """
-        Add child entity.
-
-        Args:
-            children (list)     : child information
-            metadata (dict)     : metadata
-            entity_types (list) : entity type list
-            parent_id (str)     : parent ID
-
-        Returns:
-            list: child entities
-        """
-        from rocrate.model.dataset import Dataset
-
-        entities = []
-        for child in children:
-            # Entity ID
-            if parent_id:
-                id = parent_id + '/' + child.get('name', '')
-            else:
-                id = child.get('name', '')
-
-            # Create entity
-            entity = self.crate.add(Dataset(self.crate, id, id))
-            if 'depth' in child:
-                entity['additionalType'] = entity_types[child.get('depth')]
-            if 'name' in child:
-                entity['title'] = child.get('name')
-            self.__add_properties(entity, child.get('map', {}), metadata)
-            entities.append(entity)
-
-            # Create child entity
-            child_entities = self.__add_child_entities(child.get('children', {}), metadata, entity_types, id)
-            if child_entities:
-                entity['hasPart'] = child_entities
-
-        return entities
-
-    def __get_json(self):
-        """Get RO-Crate metadata."""
-
-        return self.crate.metadata.generate()
-
-
 def create_limmiter():
     from .config import WEKO_SEARCH_UI_API_LIMIT_RATE_DEFAULT
     return Limiter(app=Flask(__name__), key_func=get_remote_address, default_limits=WEKO_SEARCH_UI_API_LIMIT_RATE_DEFAULT)
